requirements.txt
Flask==1.0.2
jieba==0.39
numpy==1.14.1
Flask_Cors==3.0.6
gunicorn==19.8.1
tensorflow-gpu==1.12.0
keras==2.2.4
gensim
PyMySQL
sshtunnel
